:truck: Move wedding components
