+++ src/components/wedding/Travel/Accommodations.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Link from '../wedding/Link';
+import Link from '../Link';
 
 export default function Accommodations() {
   return (
